requirements.txt
google-cloud-shell==1.12.1
